Changed scheduled publish coverage to use the existing test suite
ref https://linear.app/ghost/issue/ONC-1983

Keep one focused regression test for immediate duplicate-delivery responses in the existing scheduling suite.
